requirements.txt
torch
torchvision
torchaudio
efficientnet_pytorch
pillow
numpy
pandas
scikit-learn
tqdm

sentence-transformers
chromadb

openai
langchain
langchain-community

streamlit
python-dotenv

pypdf
pdfplumber

requests
beautifulsoup4
